# swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Transaction monitoring
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development server
components:
  securitySchemes:
    apikeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      # scheme: apikey
security:
  - apikeyAuth: []
...
paths:
...
  /api/rules:
    get:
      tags:
        - Rules
      summary: Get all rules
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    post:
      tags:
        - Rules
      summary: Create rule
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                name: 'Creating rule for testing'
                condition: 'Creating rule for testing'
                flag_level: 'Low'
                risk_increment: '0.8'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /api/users:
    get:
      tags:
        - Users
      summary: Get all users
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    post:
      tags:
        - Users
      summary: Create user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                name: 'Testing Create user'
                balance: '1000'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /api/evaluateRule:
    post:
      tags:
        - Evaluate rules
      summary: Evaluate rule 1 live
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                transactionIds:
                  - 04167762-3c8c-42cb-90a0-6c4c77c93422
                  - 062972c6-64a9-4519-86f9-e93899c033b9
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...